Dynamic & Temp support added for ModelData

+++ src/main/java/rs117/hd/renderer/zone/SceneUploader.java
@@ -1387,7 +1387,7 @@ class SceneUploader {
 		int x,
 		int y,
 		int z,
-		short modelOffset,
+		int modelOffset,
 		IntBuffer opaqueBuffer,
 		IntBuffer alphaBuffer
 	) {
